% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two.R
\name{add_two}
\alias{add_two}
\title{Title}
\usage{
add_two(a, b)
}
\arguments{
\item{a}{First number to add}

\item{b}{Second number to add}
}
\value{
The sum of numbers a and b
}
\description{
Title
}
\examples{
add_two(1, 4)
}
